feat: order related changes

+++ src/main/java/pizzahub/api/repositories/PizzariaRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import pizzahub.api.entities.pizzaria.Pizzaria;
 
-public interface PizzariaRepository extends JpaRepository <Pizzaria, Long> {}
+public interface PizzariaRepository extends JpaRepository <Pizzaria, Short> {}
